Move output mode unit tests to the actor suite, rather than director
