+++ tests/test_gpt4all_chat_director.py
@@ -8,7 +8,7 @@ from freezegun import freeze_time
 
 from khoj.processor.conversation import prompts
 from khoj.processor.conversation.utils import message_to_log
-from khoj.routers.helpers import aget_relevant_information_sources, aget_relevant_modes
+from khoj.routers.helpers import aget_relevant_information_sources
 from tests.helpers import ConversationFactory
 
 SKIP_TESTS = True
@@ -533,31 +533,3 @@ async def test_get_correct_tools_with_chat_history(client_offline_chat):
     # Assert
     tools = [tool.value for tool in tools]
     assert tools == ["online"]
-
-
-# ----------------------------------------------------------------------------------------------------
-@pytest.mark.anyio
-@pytest.mark.django_db(transaction=True)
-async def test_use_default_response_mode(chat_client):
-    # Arrange
-    user_query = "What's the latest in the Israel/Palestine conflict?"
-
-    # Act
-    mode = await aget_relevant_modes(user_query, {})
-
-    # Assert
-    assert mode.value == "default"
-
-
-# ----------------------------------------------------------------------------------------------------
-@pytest.mark.anyio
-@pytest.mark.django_db(transaction=True)
-async def test_use_image_response_mode(chat_client):
-    # Arrange
-    user_query = "Paint a picture of the scenery in Timbuktu in the winter"
-
-    # Act
-    mode = await aget_relevant_modes(user_query, {})
-
-    # Assert
-    assert mode.value == "image"
